small bit of refactoring
